Revert "Increase MGFXO format limits. (#7397) (fixes #7391)"
This reverts commit baab1816ef7c1c69d8922db02514af3d59133674.

+++ Tools/MonoGame.Effect.Compiler/Effect/ConstantBufferData.writer.cs
@@ -1,7 +1,5 @@
 using System.IO;
 
-using MonoGame.Framework.Utilities;
-
 namespace MonoGame.Effect
 {
     internal partial class ConstantBufferData
@@ -12,10 +10,10 @@ namespace MonoGame.Effect
 
             writer.Write((ushort)Size);
 
-            writer.Write(ParameterIndex.Count);
+            writer.Write((byte)ParameterIndex.Count);
             for (var i=0; i < ParameterIndex.Count; i++)
             {
-                writer.Write(ParameterIndex[i]);
+                writer.Write((byte)ParameterIndex[i]);
                 writer.Write((ushort)ParameterOffset[i]);
             }
         }
